JS made button colors invert when hovered over
settings.js successfully has hoverable buttons for the off and on (day and not day) buttons that change colors. style.css has a class for hoverable for the normal buttons that do not need to change color, settings.html is modified that all those buttons are part of the class hoverable. In settings.js, made new updated TO DO list. Also, changed the random number process for the ID counter for the group divs if the getDivCount returns not a number.