Changes for default styling of button; in progress blockAllUntil date:
Finished writing the blockAllUntil stuff on the settings and background
sides, just need to trouble shoot.
Created buttonOn() and buttonOff to turn the buttons on or off, removing
style changing from popup.js and settings.js.
Changed the default styling of the buttons to look like the unselected
class, so that does not need to be repeated anymore.