README update and small other changes
Popup title is on the same line as the logo.
Now using a set instead of a map in background.js.
Darker blue color for a higher contrast, and darker text for the
placeholder for a higher contrast.
README: re-wrote the top paragraph of with unique features to be a
list, added an image to the usage section, more detailed in how to
download the code in the setting up section.