Popup also updates settings page
pauseUntil, cancel pause, and block all messages also get listened to
by the settings page, with the correct changes being made so that when
the settings page changes, it does not override with the old settings.
Moved some functions to sharedFunctions.js from background.js in order
for this to work and to have the same code in both.